Fix error message in theme download
When changing to another theme and this theme needs to be downloaded
then the following message can be shown:
2021-07-31 14:45:47.892239 I  Qt: QFile::remove: Empty or null file name
This is fixed by testing if the filename is not empty.
Note that mythfrontend also incidentally crashes when changing to another
theme and this might also be fixed now.

+++ mythtv/programs/mythfrontend/themechooser.cpp
@@ -924,7 +924,9 @@ void ThemeChooser::customEvent(QEvent *e)
             m_downloadState = dsIdle;
             CloseBusyPopup();
             QStringList args = me->ExtraDataList();
-            QFile::remove(args[0]);
+
+            if (!args.isEmpty() && !args[0].isEmpty())
+                QFile::remove(args[0]);
 
             QString event = QString("THEME_INSTALLED PATH %1")
                                     .arg(m_userThemeDir +
